requirements/local.txt
Django==1.5c1
Pillow==1.7.8
South==0.7.6
django-appconf==0.5
django-ckeditor==3.6.2.1
django-compressor==1.2
django-contact-form-gv==0.4.1
django-debug-toolbar==0.9.4
-e git://github.com/ebrelsford/django-fiber.git@fa24be63859e31a9da8cd8d499950a43dcbbef0c#egg=django_fiber-dev
django-moderation==0.3.2
django-mptt==0.5.5
djangorestframework==2.1.17
docutils==0.10
geojson==1.0.1
ipython==0.13.1
johnny-cache==1.4
psycopg2==2.4.6
pytz==2012j
simplejson==3.0.7
six==1.2.0
wsgiref==0.1.2
